Fix disc splitting error
Might not be the most elegant way but didn't see another option.  Original code add the parent twice to the output template and failed to re-introduce the filename template portion that was stripped during the split.  This adds that back and corrects the second joinpath() so it references that rather than the split.  split renamed to path to indicate what it contains.